Changing all this back to using an int.
Who cares about int versus short, I'm not a professional.

+++ main/java/pokemane/northernrail/api/rail/IRailIconProvider.java
@@ -10,5 +10,5 @@ import pokemane.northernrail.api.rail.RailType;
  */
 public interface IRailIconProvider {
 	public abstract IIcon getIconFromRailType(RailType railType);
-	public abstract IIcon getIconFromRailId(short id);
+	public abstract IIcon getIconFromRailId(int id);
 }
