+++ main/java/pokemane/northernrail/api/rail/RailRegistry.java
@@ -6,14 +6,14 @@ import java.util.Hashtable;
  * Created by pokemane on 2/20/14.
  */
 public final class RailRegistry {
-	public static final Hashtable<Short, RailType> railTypeRegistry = new Hashtable<Short, RailType>();
+	public static final Hashtable<Integer, RailType> railTypeRegistry = new Hashtable<Integer, RailType>();
 
 	public static void addRailType(RailType railType){
 		if(!(railTypeRegistry.contains(railType)))
 			railTypeRegistry.put(railType.getRailId(), railType);
 	}
 
-	public static RailType getRailType(short key){
+	public static RailType getRailType(int key){
 		return railTypeRegistry.get(key);
 	}
